refactor-controller

+++ app/src/main/java/io/github/stockinventory/app/services/ProdutoService.java
@@ -5,7 +5,9 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import io.github.stockinventory.app.exceptions.ResourceNotFoundException;
 import io.github.stockinventory.app.model.Produto;
+import io.github.stockinventory.app.model.records.ProdutoRecordDTO;
 import io.github.stockinventory.app.repository.ProdutoRepository; 
 
 @Service
@@ -30,7 +32,40 @@ public class ProdutoService {
     }
 
     public void excluir(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Id do produto não pode ser encontrado");
+        }
         repository.deleteById(id);
+    }
+    public Produto updateProductById(ProdutoRecordDTO dto) {
+        if (dto.id() == null) {
+            throw new IllegalArgumentException("Id do produto não pode ser encontrado");
+        }
+
+
+        Produto entity  = repository.findById(dto.id())
+                                    .orElseThrow(() -> new ResourceNotFoundException("Produto com ID " + dto.id() + " não encontrado"));
+
+
+        boolean isUpdate = false;
+        
+        if (dto.descricao() != null && !dto.descricao().equals(entity.getDescricao())) {
+            entity.setDescricao(dto.descricao());
+            isUpdate = true;
+        }
+        if (dto.name() != null && !dto.name().equals(entity.getName())) {
+            entity.setDescricao(dto.descricao());
+            isUpdate = true;
+        }
+        if (dto.preco() != null && !dto.preco().equals(entity.getPreco())) {
+            entity.setPreco(dto.preco());
+        }
+
+        if (isUpdate) {
+            return repository.save(entity);
+        }
+
+        return entity;
     }
     
 }
